+++ RELEASES.md
@@ -27,6 +27,7 @@ Upcoming
 - Introduction of ExchangeMailListener
 - Bugfix not showing length of blob fields in "Browse a Jdbc table" in case of multiple records
 - Improve javadoc for MessageStoreSender/Listener
+- Create directory in UploadFilePipe if it doesn't exist